Add files via upload
Final product. I chose to run a counter that tracts all iterations so that I could grab the correct <open> data. This could have been done a little cleaner if I hard coded the first value an <open> value from the first <ticker> and just grabbed the rest in loop, but this was more fun.